requirements.txt
appdirs==1.4.3
apturl==0.5.2
attrs==19.3.0
Automat==0.8.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
boto==2.49.0
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.15.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
mysql-connector==2.2.9
mysql-connector-python==8.0.27
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
parsel==1.5.2
parso==0.5.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==2.0.10
protobuf==3.6.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
PyDispatcher==2.0.5
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pyxdg==0.26
PyYAML==5.3.1
queuelib==1.5.0
requests==2.22.0
requests-unixsocket==0.2.0
Scrapy==1.7.3
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
soupsieve==1.9.5
systemd-python==234
traitlets==4.3.3
Twisted==18.9.0
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
w3lib==1.21.0
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
xkit==0.0.0
zipp==1.0.0
zope.interface==4.7.1
